requirements.txt
Django>=1.4
django-tastypie
Django-facebook
South
